re-factor in preperation for conversion to using hooks
