+++ src/pages/_app.tsx
@@ -1,6 +1,5 @@
 import { AppProps } from 'next/app'
 
-import { AlertContextProvider } from '@/context/AlertContext'
 import { AuthContextProvider } from '@/context/AuthContext'
 
 import '@/styles/main.css'
@@ -9,10 +8,8 @@ const POOPS = ({ Component, pageProps }: AppProps) => {
   return (
     <>
       <AuthContextProvider>
-        <AlertContextProvider>
-          <Component {...pageProps} />
-          {/* <NavBar /> */}
-        </AlertContextProvider>
+        <Component {...pageProps} />
+        {/* <NavBar /> */}
       </AuthContextProvider>
     </>
   )
